no srp is respected in this household

+++ src/main/java/com/uq/jokievents/model/LocalityOrder.java
@@ -19,5 +19,5 @@ public class LocalityOrder {
 
     private int numTicketsSelected;
     private String localityName;
-    private float totalPaymentAmount;
+    private Double totalPaymentAmount;
 }
